move cmake_prefix_path to config script
